pyproject.toml
[tool.poetry]
name = "scrabbleScoreboard"
version = "0.1.0-beta"
description = "Scrabble scoreboard, analytics and utilities"
authors = ["Andrea Rondón <andrea.estefania.rv@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7, <4"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.4.4"
flask-marshmallow = "^0.14.0"
Flask-Migrate = "^2.5.3"
Flask-JWT-Extended = "^3.25.0"
python-dotenv = "^0.15.0"
setuptools = "^51.1.2"
flasgger = "^0.9.5"
marshmallow-enum = "^1.5.1"
marshmallow-sqlalchemy = "^0.24.1"
click = "^7.1.2"
Flask-Admin = "^1.5.7"
WTForms-SQLAlchemy = "^0.2"
Flask-Login = "^0.5.0"
Flask-WTF = "^0.14.3"
Flask-Bootstrap = "^3.3.7"
Flask-Caching = "^1.9.0"
webargs = "^7.0.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.1.2"
tox = "^3.21.2"
pytest-factoryboy = "^2.1.0"
pytest-flask = "^1.1.0"
pytest-runner = "^5.2"
passlib = "^1.7.4"
pytest-rerunfailures = "^9.1.1"
coverage = "^5.4"
typing-extensions = "^3.7.4"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
